chore: Clarify process for API WG approval of RFCs (#23)
Our RFC approval process has required some tweaks over time, because
presenting to a quorum of the WG has proved challenging due to member
timezones and engagement levels.

This PR documents a simple majority of API WG sign-offs as the
requirement for approval, and describes the presentation of the RFC in a
synchronous meeting as an optional step when asynchronous PR feedback is
insufficient.